requirements.txt
# PyTorch with CPU optimizations
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.1.1

# Core dependencies
streamlit>=1.31.0
yt-dlp>=2023.11.16
openai-whisper>=20231117
requests>=2.31.0
tweepy>=4.14.0
emoji>=2.8.0
python-dotenv>=1.0.0

# Streamlit enhancements
watchdog>=3.0.0  # For better file handling in Streamlit
streamlit-extras>=0.3.6  # Additional Streamlit components

# --- Celery background job runner dependencies ---
celery>=5.3
redis>=4.5

# --- Testing dependencies ---
pytest>=8.0.0  # Testing framework
pytest-asyncio>=0.23.5  # Async test support
pytest-cov>=4.1.0  # Coverage reporting
pytest-mock>=3.12.0  # Mocking support

# --- Type checking and stubs ---
mypy>=1.8.0
types-requests>=2.31.0
types-redis>=4.6.0
types-emoji>=2.1.0.3  # Latest available version